process_records() not quite used in add, so moved back under update.update; Cosmetic changes to update.update_otu
